Adding thumbnail_url() to cropduster fields as a convenience and for backwards compat
Not having this method was going to require a slew of changes throughout our
codebase, but moreover, it seems like it's actually a helper worth having;
please let me know if I have missed out on an existing way of doing this that's
equally concise and cropduster-idiomatic.

If this gets accepted upstream, folks might consider renaming the method; just
leaving it with this name because that's what we had in old-cropduster.